provide a new api: render, which takes advantage of horizon's internal state
