+++ demo/simple.html
@@ -271,8 +271,8 @@
     var foo = random("foo"),
         bar = random("bar");
 
-    console.log(foo.add(bar));
-    console.log(foo.subtract(bar));
+    // console.log(foo.add(bar));
+    // console.log(foo.subtract(bar));
 
     d3.select("#example1").call(function(div) {
 
@@ -288,7 +288,8 @@
                 .enter().append("div")
             .attr("class", "horizon");
 
-        context.horizon(h1).extent([-20, 20])
+        const h0 = context.horizon().height(120).extent([-20, 20]);
+        h0.render(h1);
             // .call(context.horizon()._extent([-20, 20]));
 
         div.append("div")
@@ -307,22 +308,25 @@
         const a = div.append("div")
             .attr("class", "horizon");
 
-        context.horizon(a)
+        const h = context.horizon()
             .height(120)
             .mode("mirror")
             .colors(["#bdd7e7","#bae4b3"])
             .title("Area (120px)")
             .extent([-10, 10]);
+        h.render(a);
 
         const b = div.append("div")
             .attr("class", "horizon");
 
-        context.horizon(b)
-            .height(30)
+        const h2 = context.horizon()
+            .height(80)
             .mode("mirror")
             .colors(["#bdd7e7","#bae4b3"])
             .title("Area (30px)")
             .extent([-10, 10]);
+
+        h2.render(b);
     });
 
     d3.select("#example2b").call(function(div) {
@@ -331,38 +335,42 @@
         const c = div.append("div")
             .attr("class", "horizon");
 
-        context.horizon(c)
+        const h3 = context.horizon()
             .height(120)
             .colors(["#bdd7e7","#bae4b3"])
             .title("Horizon, 1-band (120px)")
-            .extent([-10, 10])
+            .extent([-10, 10]);
+        h3.render(c);
 
         const d = div.append("div")
             .attr("class", "horizon")
-        context.horizon(d)
+        const h5 = context.horizon()
             .height(60)
             .colors(["#6baed6","#bdd7e7","#bae4b3","#74c476"])
             .title("Horizon, 2-band (60px)")
-            .extent([-10, 10])
+            .extent([-10, 10]);
+        h5.render(c);
 
         const e = div.append("div")
             .attr("class", "horizon");
 
 
-        context.horizon(e)
-            .height(40)
+        const h6  = context.horizon()
+            .height(120)
             .colors(["#3182bd","#6baed6","#bdd7e7","#bae4b3","#74c476","#31a354"])
             .title("Horizon, 3-band (40px)")
             .extent([-10, 10])
+        h6.render(c);
 
         const f = div.append("div")
             .attr("class", "horizon");
 
-        context.horizon(f)
+        const h7 = context.horizon()
             .height(30)
             .colors(["#08519c","#3182bd","#6baed6","#bdd7e7","#bae4b3","#74c476","#31a354","#006d2c"])
             .title("Horizon, 4-band (30px)")
-            .extent([-10, 10])
+            .extent([-10, 10]);
+        h7.render(f);
 
     });
